% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expand_Rscript}
\alias{expand_Rscript}
\title{File path handler}
\usage{
expand_Rscript(fps, debug = F)
}
\arguments{
\item{fps}{A character vector of paths. The vector can be a mix of either
(1) directory paths, (2) file paths, or (3) glob expressions to
R scripts. Note that (2) and (3) must end with \code{.R} or \code{.r}.
See \code{\link[=expand_Rscript]{expand_Rscript()}} for details.}

\item{debug}{Logical. Whether to return results as a list with \code{fps}
as names.}
}
\value{
A character vector or a list.
}
\description{
File path handler
}
\examples{
expand_Rscript = fundoc:::expand_Rscript
dir_ = system.file(package = "fundoc")
(a_file = file.path(dir_, "foo.R"))
expand_Rscript(a_file)
(a_dir = file.path(dir_, "R"))
expand_Rscript(a_dir)
(a_glob = file.path(dir_, "*", "*.R"))
expand_Rscript(a_glob)
# Mix file paths
expand_Rscript( c(a_file, a_dir, a_glob) )
expand_Rscript( c(a_file, a_dir, a_glob), debug = T )
}
