% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundoc-package.R
\docType{package}
\name{fundoc-package}
\alias{fundoc-package}
\title{fundoc: Quick and dirty function documentation}
\description{
Generate function documentation along with evaluated examples from R scripts documented in Roxygen comments
}
\author{
\strong{Maintainer}: Yongfu Liao \email{liao961120@gmail.com} (\href{https://orcid.org/0000-0002-1814-2993}{ORCID})

}
\keyword{internal}
