% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{locate_description}
\alias{locate_description}
\title{Locate DESCRIPTION content from a bunch of files}
\usage{
locate_description(files)
}
\arguments{
\item{files}{A character vector of file paths.}
}
\description{
Given a vector of file paths, this function tries to find
and parse the R script with DESCRIPTION content written as Roxygen
docstring at the beginning of the file. If more than one
file contain DESCRIPTION content, only the first one would
be parsed and returned.
}
