% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_description_in_script}
\alias{parse_description_in_script}
\title{Extract embedded DESCRIPTION content in an R script}
\usage{
parse_description_in_script(fp)
}
\description{
Extract embedded DESCRIPTION content in an R script
}
\examples{
fp = system.file("foo.R", package="fundoc")
xfun::file_string(fp)
fundoc:::parse_description_in_script(fp) |> cat()
}
